+++ test/Libplanet.Net.Tests/SwarmTest.Broadcast.cs
@@ -924,7 +924,7 @@ namespace Libplanet.Net.Tests
                 await BootstrapAsync(swarmC, swarmA.AsPeer);
 
                 await swarmC.PullBlocksAsync(
-                    TimeSpan.FromSeconds(5), int.MaxValue, int.MaxValue, null, default);
+                    TimeSpan.FromSeconds(5), int.MaxValue, null, default);
                 await swarmC.BlockAppended.WaitAsync();
                 Assert.Equal(chainC.Tip, chainATip);
             }